---- Version 1 PiedPiper2 ----
This version allow you to play basically to the game.
You can choose a number of ships, a team, a ship for your team.
When the game start, just one ship is different meanings that it's the intruder you have to touch on to win.
Sounds and vibrations happen when touching on the screen.
When winning, the ship change to an explosion and a popup window shows you that you win to finally get you back on the menu.

+++ app/src/main/java/com/example/herouard/piedpiper/PopUp.java
@@ -33,7 +33,7 @@ public class PopUp extends Activity {
         int width = dm.widthPixels;
         int height = dm.heightPixels;
 
-        getWindow().setLayout((int)(width),(int)(height));
+        getWindow().setLayout(width, height);
         getWindow().setBackgroundDrawableResource(R.color.translucent_black);
         ll2.setBackgroundResource(R.drawable.heart);
